feat: batchMint recurring function
